requirements.txt
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf>=5.26.1
scikit-learn==1.3.2